\name{RcppDocs}
\alias{RcppDocs}
\alias{showRcppDoc}
\alias{showRcppQuickRef}
\title{Displays Documentation on Rcpp Class Library.}
\description{
  Displays either the Rcpp library PDF file or the Rcpp Quick Reference
  txt file.
  }
  \usage{
showRcppDoc()
showRcppQuickRef()
}
\author{Dominick Samperi}

