\name{unitTests}
\alias{seqStrict}
\alias{unitTests}
\alias{runcxxPackTests}
\title{Runs unit tests for cxxPack}
\description{
  Runs unit tests for the \code{cxxPack} package.
}
\usage{
runcxxPackTests()
}
\details{
  Currently the unit tests exercise the \code{serialNumber} function.
}
\author{Dominick Samperi}

\keyword{misc}
